requirements.txt
Django==1.7.1
Pillow==2.6.1
django-appconf==0.6
django-bower==5.0.1
django-cors-headers==1.0.0
django-custom-user==0.5
django-daterange-filter==1.0.0
django-grappelli==2.6.3
django-location-field==1.5.2
django-sitetree==1.2.0
django-tastypie==0.12.1
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
sorl-thumbnail==11.12.1b
django-simple-email-confirmation==0.12
django-summernote==0.6.2